fixed error with activation (re)added the toggle function to schedules added rudimentary acl to non-department dependant views removed user views

+++ com_thm_organizer/admin/views/monitor_manager/view.html.php
@@ -47,10 +47,17 @@ class THM_OrganizerViewMonitor_Manager extends THM_CoreViewList
     protected function addToolBar()
     {
         JToolbarHelper::title(JText::_('COM_THM_ORGANIZER_MONITOR_MANAGER_VIEW_TITLE'), 'organizer_monitors');
-        JToolbarHelper::addNew('monitor.add');
-        JToolbarHelper::editList('monitor.edit');
-        JToolbarHelper::deleteList(JText::_('COM_THM_ORGANIZER_ACTION _DELETE_CONFIRM'), 'monitor.delete');
-        JToolbarHelper::divider();
-        JToolbarHelper::preferences('com_thm_organizer');
+
+        $actions = $this->getModel()->actions;
+
+        $isAdmin = ($actions->{'core.admin'});
+        if ($isAdmin)
+        {
+            JToolbarHelper::addNew('monitor.add');
+            JToolbarHelper::editList('monitor.edit');
+            JToolbarHelper::deleteList(JText::_('COM_THM_ORGANIZER_ACTION _DELETE_CONFIRM'), 'monitor.delete');
+            JToolbarHelper::divider();
+            JToolbarHelper::preferences('com_thm_organizer');
+        }
     }
 }
